requirements.txt
-e git+https://github.com/openskynetwork/opensky-api.git#egg=opensky-api&subdirectory=python
cryptography
pandas
pyyaml
click
structlog

pytest
pytest-cov
freezegun
requests-mock
